+++ src/components/TwicImg.tsx
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
-import { View, Image, Text, StyleSheet } from 'react-native';
+import { Animated, Image, Platform, StyleSheet, Text, View } from 'react-native';
 import {
     parseAlt,
     parseAnchor,
     parseFocus,
     parseMode,
+    parsePlaceholder,
     parsePreTransform,
     parseRatio,
     parseSrc,
@@ -24,22 +25,37 @@ export interface WrapperAttributes extends Attributes {
     width: number;
     height: number;
 }
-const TwicWrapper = (props: WrapperAttributes) => {
+const TwicWrapper = React.memo((props: WrapperAttributes) => {
     let width, height;
     ({ width, height } = props);
     const alt = parseAlt(props.alt);
     const anchor = parseAnchor(props.anchor);
     const focus = parseFocus(props.focus);
     const mode = parseMode(props.mode);
+    const placeholder = parsePlaceholder(props.placeholder);
     const preTransform = parsePreTransform(props.preTransform);
     const ratio = parseRatio(props.ratio);
     const src = parseSrc(props.src);
     const size = computeSize(width, height, ratio);
     const step = parseStep(props.step);
     const computedSrc =
-        size.width && computeSrc(anchor, focus, mode, preTransform, size, src, step);
-    console.log('computedSrc', computedSrc);
+        size.width && computeSrc(anchor, focus, mode, undefined, preTransform, size, src, step);
+    let computedPlaceholderSrc =
+        placeholder === `preview` &&
+        computeSrc(anchor, focus, mode, placeholder, preTransform, size, src, step);
+    const imageTransition = new Animated.Value(placeholder ? 1 : 0);
+    const onImage = () => {
+        Animated.timing(imageTransition, {
+            toValue: 0,
+            duration: 200,
+            useNativeDriver: true
+        }).start();
+    };
+    const blur = () => (Platform.OS === `web` ? 0 : 5);
     if (computedSrc) {
+        if (config.debug) {
+            console.debug('Computed src ', computedSrc, computedPlaceholderSrc);
+        }
         return (
             <View style={computeWrapperStyle(size, styles)}>
                 <Image
@@ -47,7 +63,17 @@ const TwicWrapper = (props: WrapperAttributes) => {
                     source={{ uri: computedSrc }}
                     resizeMode={mode}
                     style={styles.img}
+                    onLoad={onImage}
                 />
+                {computedPlaceholderSrc && (
+                    <Animated.Image
+                        accessibilityLabel={computeAlt(alt, src)}
+                        blurRadius={blur()}
+                        source={{ uri: computedPlaceholderSrc }}
+                        resizeMode={mode}
+                        style={[styles.placeholdeImg, { opacity: imageTransition }]}
+                    />
+                )}
                 {config.debug && (
                     <View style={styles.debug}>
                         <Text>
@@ -60,7 +86,8 @@ const TwicWrapper = (props: WrapperAttributes) => {
     } else {
         return null;
     }
-};
+});
+
 export default class TwicImg extends Component<Attributes, WrapperState> {
     constructor(props: Attributes) {
         super(props);
@@ -69,6 +96,7 @@ export default class TwicImg extends Component<Attributes, WrapperState> {
             height: 0
         };
     }
+
     render() {
         const { props } = this;
         const ratio = parseRatio(props.ratio);
@@ -78,7 +106,9 @@ export default class TwicImg extends Component<Attributes, WrapperState> {
                 style={StyleSheet.flatten([styles.layout, style])}
                 onLayout={(event) => {
                     let { x, y, width, height } = event.nativeEvent.layout;
-                    console.log('onLayout', x, y, width, height);
+                    if (config.debug) {
+                        console.debug('onLayout', x, y, width, height);
+                    }
                     this.setState({
                         width: width,
                         height: height
@@ -97,16 +127,23 @@ const styles = StyleSheet.create({
         position: 'absolute',
         backgroundColor: '#cccccc'
     },
-    img: {
-        width: '100%',
-        height: '100%',
+    placeholdeImg: {
         position: 'absolute',
-        top: 0
+        top: 0,
+        bottom: 0,
+        left: 0,
+        right: 0
+    },
+    img: {
+        position: 'absolute',
+        top: 0,
+        bottom: 0,
+        left: 0,
+        right: 0
     },
     layout: {
         overflow: `hidden`,
-        width: '100%',
-        backgroundColor: '#FF0000'
+        width: '100%'
     },
     wrapper: {
         overflow: `hidden`,